+++ releasenotes.md
@@ -6,6 +6,7 @@
 - Lorem.ParagraphByChars(int chars = 256, bool supplemental = false)
 - Lorem.Paragraphs(int paragraphCount = 3, bool supplemental = false): Optional parameters and default values added.
 - Lorem.Question(int wordCount = 4, bool supplemental = false, int randomWordsToAdd = 0)
+- Lorem.Questions(int questionCount = 3, bool supplemental = false)
 
 ## Version 0.8.0
 ### New Features